# models/product.go
package models

import "time"

type Product struct {
	ID        uint       `json:"id" gorm:"primaryKey"`
	CreatedAt time.Time  `json:"created_at"`
	UpdatedAt time.Time  `json:"updated_at"`
	DeletedAt *time.Time `json:"deleted_at" gorm:"index"`

	Name  string  `json:"name"`
	Price float64 `json:"price"`
	Stock int     `json:"stock"`
}